Support heapallindexed check in GPDB.
Implement necessary gymnastics.
Extend IndexBuildScan's callback to accept HeapTuple instead of
ctid.
