+++ contrib/amcheck/amcheck--1.0--1.1.sql
@@ -13,7 +13,7 @@ CREATE FUNCTION bt_index_check(index regclass,
     heapallindexed boolean)
 RETURNS VOID
 AS 'MODULE_PATHNAME', 'bt_index_check'
-LANGUAGE C STRICT PARALLEL RESTRICTED;
+LANGUAGE C STRICT;
 
 --
 -- bt_index_parent_check()
@@ -22,7 +22,7 @@ CREATE FUNCTION bt_index_parent_check(index regclass,
     heapallindexed boolean)
 RETURNS VOID
 AS 'MODULE_PATHNAME', 'bt_index_parent_check'
-LANGUAGE C STRICT PARALLEL RESTRICTED;
+LANGUAGE C STRICT;
 
 -- Don't want these to be available to public
 REVOKE ALL ON FUNCTION bt_index_check(regclass, boolean) FROM PUBLIC;
