+++ contrib/amcheck/amcheck.control
@@ -1,5 +1,5 @@
 # amcheck extension
 comment = 'functions for verifying relation integrity'
-default_version = '1.1'
+default_version = '1.0'
 module_pathname = '$libdir/amcheck'
 relocatable = true
