+++ contrib/amcheck/expected/check.out
@@ -1,1 +1,2 @@
 CREATE EXTENSION amcheck;
+ALTER EXTENSION amcheck UPDATE TO '1.1';
