+++ contrib/amcheck/sql/check.sql
@@ -1,1 +1,2 @@
 CREATE EXTENSION amcheck;
+ALTER EXTENSION amcheck UPDATE TO '1.1';
