+++ src/include/catalog/index.h
@@ -139,4 +139,15 @@ extern Oid	IndexGetRelation(Oid indexId, bool missing_ok);
 
 extern void IndexSetParentIndex(Relation idx, Oid parentOid);
 
+extern double
+IndexBuildHeapScan(Relation heapRelation,
+				   Relation indexRelation,
+				   struct IndexInfo *indexInfo,
+				   bool allow_sync,
+				   EState *estate,
+				   Snapshot snapshot,
+				   TransactionId OldestXmin,
+				   IndexBuildCallback callback,
+				   void *callback_state);
+
 #endif   /* INDEX_H */
